requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
google-generativeai>=0.3.2
PyPDF2>=3.0.1
python-docx>=1.1.0
pydantic-settings>=2.1.0
httpx>=0.25.2
jinja2>=3.1.2